refactor: compute_overrides
